[#122] fix: 0표인 종료 투표의 투표 버튼 텍스트 버튼 색상 수정
